requirements.txt
#pysqlite3-binary==0.5.2
chromadb==0.4.13
llama-cpp-python==0.1.78
langchain==0.0.302
sentence-transformers==2.2.2
flask==2.3.2
flask-sock==0.6.0
pypdf==3.11.0
python-dotenv==1.0.0
atlassian-python-api==3.39.0
beautifulsoup4==4.12.2
lxml==4.9.2
cryptography==40.0.2
alt-profanity-check==1.3.0
gunicorn==21.2.0
markdownify==0.11.6
flask-cors==4.0.0